Fix (ASP): Log `php` errors to `stderr` when running in `docker`
Previously, `ASP` `php` errors were logged to a file (`/src/ASP/system/logs/php_errors.log`). When running in docker, logging to a file can trip up users of `docker` who expect errors to be logged to `stderr`.

Now, the env variable `PHP_VERSION` (present in official `php` images) is used to detect whether `ASP` is running in `docker`. If `ASP` is running in `docker`, `php` errors are logged to `stderr`. Otherwise, `php` errors are logged to a file.