Fix (ASP): Fix `Serverinfo` module to use gamespy query instead of Rcon to determine server online status
Previously, the ASP `ServerInfo` module used Rcon against the BF2 server to determine server status. However, the Rcon port may not always be available (i.e. not publically accessible). In contrast, the BF2 server gamespy port is always available to the public (to be listed on a master server) and is a better way to determine server status.

Now, a gamespy query is used to determine BF2 server status instead. If there is no response from the server, it is assumed to be offline.