Fix (ASP): Speed up `Serverinfo` module determination of server status by avoiding hitting socket timeout when possible
Previously, 0.5s timeout is hit regardless of whether there is a gamespy query response from the server. However, to speed up the status checks, timeout should only be hit if the server does not respond.

Send the gamespy query, and wait for a response. If there is a response, validate the first 5 byte and the server is online. If there is no response for 0.5s, assume the server is offline.

Follow-up of #104