Feature: Add `publicaddress` column in `servers` table and ability to manage it in `ASP` `Serverinfo` module
The new `publicaddress` column of the `servers` table may be a DNS or IP address, and is useful for display purposes instead of the `ip` column especially when the `ip` is a private IP.

The ASP `Serverinfo` module table now supports AJAX hydration. The `Set Rcon Data` button of each server is renamed to `Manage`, and now allows editing of the `publicaddress` of a server.