Enhancement: Increase `prefix` column length in `servers` table to `100` characters
Previously,  if `snapshot_prefix` is longer than `30` characters in `BF2StatisticsConfig.py`, the server will not be `INSERT`ed into the DB by `bf2statistics.php` when it processes the server's snapshot at the end of a round.

Now, server prefixes length are increased to `100` characters. This allows for longer prefixes in `BF2StatisticsConfig.py`.